API/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
flask-cors==4.0.1
flask-mysqldb==2.0.0
Flask-Session
python-dateutil